# ml_service/openapi.json
{"openapi":"3.1.0","info":{"title":"Moscow tram forecast","description":"Preserved public-best forecast (0.89402), Nov–Dec 2025. Day/month slicing; no live forecasting or stop-level target estimates.","version":"1.0.0"},"paths":{"/healthz":{"get":{"summary":"Health","operationId":"health_healthz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/readyz":{"get":{"summary":"Ready","operationId":"ready_readyz_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/model":{"get":{"summary":"Model","operationId":"model_v1_model_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/sources":{"get":{"summary":"Sources","operationId":"sources_v1_sources_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/routes":{"get":{"summary":"Routes","operationId":"routes_v1_routes_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/v1/geo/{kind}":{"get":{"summary":"Geo","operationId":"geo_v1_geo__kind__get","parameters":[{"name":"kind","in":"path","required":true,"schema":{"enum":["routes","stops"],"type":"string","title":"Kind"}},{"name":"route","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Route"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/forecast":{"get":{"summary":"Forecast","operationId":"forecast_v1_forecast_get","parameters":[{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2025-11-01","title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2025-12-31","title":"End"}},{"name":"routes","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"integer"},"maxItems":10},{"type":"null"}],"title":"Routes"}},{"name":"aggregation","in":"query","required":false,"schema":{"enum":["hour","day","month","total"],"type":"string","default":"hour","title":"Aggregation"}},{"name":"hour_from","in":"query","required":false,"schema":{"type":"integer","maximum":23,"minimum":0,"default":0,"title":"Hour From"}},{"name":"hour_to","in":"query","required":false,"schema":{"type":"integer","maximum":24,"minimum":1,"default":24,"title":"Hour To"}},{"name":"weather_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Weather Factor"}},{"name":"event_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Event Factor"}},{"name":"season_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Season Factor"}},{"name":"stop_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stop Id"}},{"name":"segment_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Segment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/v1/forecast.csv":{"get":{"summary":"Export","operationId":"export_v1_forecast_csv_get","parameters":[{"name":"start","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2025-11-01","title":"Start"}},{"name":"end","in":"query","required":false,"schema":{"type":"string","format":"date","default":"2025-12-31","title":"End"}},{"name":"routes","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"integer"},"maxItems":10},{"type":"null"}],"title":"Routes"}},{"name":"aggregation","in":"query","required":false,"schema":{"enum":["hour","day","month","total"],"type":"string","default":"hour","title":"Aggregation"}},{"name":"hour_from","in":"query","required":false,"schema":{"type":"integer","maximum":23,"minimum":0,"default":0,"title":"Hour From"}},{"name":"hour_to","in":"query","required":false,"schema":{"type":"integer","maximum":24,"minimum":1,"default":24,"title":"Hour To"}},{"name":"weather_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Weather Factor"}},{"name":"event_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Event Factor"}},{"name":"season_factor","in":"query","required":false,"schema":{"type":"number","maximum":3,"minimum":0,"default":1,"title":"Season Factor"}},{"name":"stop_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stop Id"}},{"name":"segment_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Segment Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}